Changed formatting to be like 2024 Code (one swerve module file and one swerve drive file)
