+++ .vscode/settings.json
@@ -1,4 +1,9 @@
 {
+  "java.project.referencedLibraries": [
+    "lib/**/*.jar",
+    "c:\\path\\to\\jarfile\\commons-logging-1.1.1.jar"
+  ],
+  
   "java.configuration.updateBuildConfiguration": "automatic",
   "java.server.launchMode": "Standard",
   "files.exclude": {
