+++ src/main/java/frc/robot/RobotContainer.java
@@ -33,7 +33,7 @@ public class RobotContainer {
   private final ExampleCommand exampleCommand = new ExampleCommand();
   
   public RobotContainer() {
-    drive = SwerveDrive.create();
+    drive = new SwerveDrive();
 
     field = new FieldSim(drive);
